Search sequence embedded in navbar
